feat(apps): enhance online template deployment with dynamic config

+++ backend/api/db/schemas/apps.py
@@ -51,7 +51,7 @@ class DeployForm(BaseModel):
     cap_add: Optional[List[str]]
     network_mode: Optional[str]
     network: Optional[str]
-    cpus: Optional[int]
+    cpus: Optional[float]
     mem_limit: Optional[str]
     edit: Optional[bool]
     id: Optional[str]
